+++ Dockerfile.upload_failures
@@ -1,9 +1,9 @@
 FROM public.ecr.aws/lambda/python:3.8
 
-COPY backend/corpora/upload_failures/requirements.txt requirements.txt
+COPY backend/upload_failures/requirements.txt requirements.txt
 RUN pip3 install -r requirements.txt
 
-COPY backend/corpora/upload_failures .
+COPY backend/upload_failures .
 COPY backend/__init__.py ./backend/__init__.py
 COPY backend/corpora/__init__.py ./backend/corpora/__init__.py
 COPY backend/common ./backend/common
